+++ extensions/slack/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Slack Changelog
 
+## [Fix YAML manifest errors in README] - 2025-06-25
+
+## [Fix search messages from specific user] - 2025-06-23
+
+- Fix search messages from specific user (using `from:<@${user.id}>` instead of `from:${user.name}`)
+
 ## [Fix Emoji Search Missing Scope Error] - 2025-05-28
 
 - Add missing `emoji:read` scope to fix "missing_scope" error when using Search Emojis command